Refactored to be Clean Architecture. Need to fix tests
