+++ README-RU.md
@@ -46,7 +46,7 @@ make BINARIES_PATH=. build-current
 
 ## Запуск тестов
 
-Есть тесты для [internal/json_storage](internal/json_storage/json_storage_test.go) и [internal/usecases](internal/usecases)
+Есть тесты для [internal/json_storage](internal/json_storage/json_storage_test.go) и [internal/usecases](internal/core)
 ```shell
 make test
 make test-race
