--- internal/console/common.go
@@ -1,18 +1,0 @@
-package console
-
-import (
-	"io"
-)
-
-type Command interface {
-	GetReader() io.Reader
-	GetWriter() io.Writer
-	Done() <-chan error
-	Kill() error
-}
-
-type Options struct {
-	Cols uint16
-	Rows uint16
-	Env  []string
-}
